CachedResource -> ResourceResponse, Resource -> CachedResource
