+++ lib/cached_resource.js
@@ -1,98 +1,75 @@
-const LOCK_TIMEOUT = 30 * 1000; // 30 seconds
+const MAX_RESOURCES_CACHE = 50;
+const MAX_RESOURCE_AGE = 1000 * 60 * 60 * 24; // 24 hours
+const DEFAULT_RESOURCE_FRESHNESS = 60 * 1000; // 60 seconds
+const SUCCESSFUL_RESPONSE_CODES = [200];
 
-var CachedResource = function(spec) {
-  this.data = spec.data;
-  this.expires_at = expiresAt(spec.response, spec.request_time, spec.response_time);
-  this.locked_at = 0;
+var _ = require('underscore');
+var lru = require('lru-cache');
+var Resource = require('solidus-client/lib/resource');
+var ResourceResponse = require('./resource_response.js');
 
-  if (this.expires_at) {
-    this.has_expiration = true;
+var cache = lru({
+  max: MAX_RESOURCES_CACHE,
+  length: function() {return 1},
+  maxAge: MAX_RESOURCE_AGE
+});
+
+var CachedResource = function(options, auth, params, logger) {
+  this.resource = new Resource(options, auth, params);
+  this.logger = logger;
+};
+
+CachedResource.prototype.get = function(callback) {
+  if (!this.resource.url) return callback(null, {});
+
+  // TODO: include the auth in the cache key
+  var resource_response = cache.get(this.resource.url);
+  if (resource_response) {
+    this.logger.log('Resource recovered from cache: ' + this.resource.url, 3);
+    refreshCache.call(this, resource_response);
+    callback(null, resource_response);
   } else {
-    this.has_expiration = false;
-    this.expires_at = new Date().getTime();
+    getAndCache.call(this, callback);
   }
+};
 
-  this.maxAge = function() {
-    return Math.max(0, Math.ceil((this.expires_at - new Date().getTime()) / 1000));
-  };
+// PRIVATE
 
-  this.expired = function() {
-    return new Date().getTime() >= this.expires_at;
-  };
+var refreshCache = function(resource_response) {
+  if (!resource_response.expired() || !resource_response.lock()) return;
 
-  this.lock = function() {
-    if (new Date().getTime() >= this.locked_at + LOCK_TIMEOUT) {
-      this.locked_at = new Date().getTime();
-      return true;
-    } else {
-      return false;
+  // The cache is expired, refresh it in the next event loop cycle
+  var self = this;
+  process.nextTick(function() {
+    self.logger.log('Refreshing expired cached resource: ' + self.resource.url, 3);
+    getAndCache.call(self, function() {
+      resource_response.unlock();
+    });
+  });
+};
+
+var getAndCache = function(callback) {
+  var self = this;
+
+  self.resource.get(function(err, res) {
+    if (err) return callback(err, null);
+
+    self.logger.log('Requested resource: [' + res.response.statusCode + '] ' + self.resource.url, 3);
+    var resource_response = new ResourceResponse(res);
+
+    if (SUCCESSFUL_RESPONSE_CODES.indexOf(res.response.statusCode) > -1) {
+      if (!resource_response.has_expiration) {
+        // The response has no cache headers, cache for a default duration
+        resource_response.expires_at = new Date().getTime() + DEFAULT_RESOURCE_FRESHNESS;
+      }
+      // TODO: include the auth in the cache key
+      cache.set(self.resource.url, resource_response);
     }
-  };
 
-  this.unlock = function() {
-    this.locked_at = 0;
-  };
+    callback(null, resource_response);
+  });
 };
 
-// https://developer.mozilla.org/en/docs/HTTP_Caching_FAQ
-var expiresAt = function(response, request_time, response_time) {
-  var freshness_lifetime = freshnessLifetime(response, request_time);
-  if (freshness_lifetime === null) return null;
-
-  return new Date().getTime() + freshness_lifetime - currentAge(response, request_time, response_time);
-};
-
-// http://tools.ietf.org/html/rfc2616#section-13.2.4
-var freshnessLifetime = function(response, request_time) {
-  var cache_control = parseCacheControl(response.headers['cache-control'] || '');
-
-  if (cache_control.hasOwnProperty('s-maxage') && cache_control['s-maxage'] !== true) {
-    return cache_control['s-maxage'] * 1000;
-  }
-
-  if (cache_control.hasOwnProperty('max-age') && cache_control['max-age'] !== true) {
-    return cache_control['max-age'] * 1000;
-  }
-
-  if (response.headers['expires']) {
-    var expires = parseDate(response.headers['expires']);
-    var date = parseDate(response.headers['date']) || request_time;
-    return expires - date;
-  }
-
-  return null;
-};
-
-// http://tools.ietf.org/html/rfc2616#section-13.2.3
-// http://greenbytes.de/tech/webdav/draft-ietf-httpbis-p6-cache-22.html#age.calculations
-var currentAge = function(response, request_time, response_time) {
-  var apparent_age = Math.max(0, response_time - (parseDate(response.headers['date']) || request_time));
-  var response_delay = response_time - request_time;
-  var corrected_age_value = response.headers['age'] ? response.headers['age'] * 1000 + response_delay : 0;
-  var corrected_initial_age = Math.max(apparent_age, corrected_age_value);
-  var resident_time = new Date().getTime() - response_time;
-  var current_age = corrected_initial_age + resident_time;
-
-  return current_age;
-};
-
-var parseCacheControl = function(header) {
-  var directives = header.split(',');
-  var obj = {};
-
-  for (var i = 0, len = directives.length; i < len; i++) {
-    var parts = directives[i].split('=');
-    var key = parts.shift().trim();
-    var val = parseInt(parts.shift(), 10);
-
-    obj[key] = isNaN(val) ? true : val;
-  }
-
-  return obj;
-};
-
-var parseDate = function(header) {
-  return header ? Date.parse(header) : null;
-};
+CachedResource.cache = cache;
 
 module.exports = CachedResource;
