+++ lib/page.js
@@ -12,7 +12,7 @@ var EventEmitter = require('events').EventEmitter;
 var _ = require('underscore');
 var async = require('async');
 
-var Resource = require('./resource.js');
+var CachedResource = require('./cached_resource.js');
 
 // rounds datetime to nearest 5 minutes (in the past)
 var getRoundedTime = function( datetime, round_by ){
@@ -168,16 +168,16 @@ var Page = function( page_path, options ){
     if( page.params.resources ){
       // convert resources object into array
       var resources_array = _( page.params.resources ).map( function( options, name ){
-        var resource = new Resource(options, server.auth, params, server.logger);
+        var resource = new CachedResource(options, server.auth, params, server.logger);
         resource.name = name;
         return resource;
       });
       async.each( resources_array, function( resource, cb ){
-        resource.get(function(err, cached_resource) {
+        resource.get(function(err, resource_response) {
           if( err ){
             server.logger.log( 'Error retrieving resource "'+ resource.name +'": '+ err, 3 );
           } else {
-            iterator(resource, cached_resource);
+            iterator(resource, resource_response);
           }
           cb();
         });
@@ -266,8 +266,8 @@ var Page = function( page_path, options ){
 
     var start_resources = new Date;
     this.fetchResources( context.parameters,
-      function(resource, cached_resource) {
-        context.resources[resource.name] = cached_resource.data;
+      function(resource, resource_response) {
+        context.resources[resource.name] = resource_response.data;
       },
       function() {
         server.logger.log( page.route +' resources fetched in '+ ( new Date - start_resources ) +'ms', 3 );
