+++ lib/server.js
@@ -39,7 +39,7 @@ var Page = require('./page.js');
 var Preprocessor = require('./preprocessor.js');
 var Redirect = require('./redirect.js');
 var Logger = require('./logger.js');
-var Resource = require('./resource.js');
+var CachedResource = require('./cached_resource.js');
 
 // make the path into a Windows compatible path
 var deGlobifyPath = function( file_path ){
@@ -298,19 +298,19 @@ var SolidusServer = function( options ){
     var self = this;
     this.router.get(this.options.api_route + 'resource.json', function(req, res) {
       if (!req.query.url) return res.json(400, {error: "Missing 'url' parameter"});
-      var resource = new Resource(req.query.url, solidus_server.auth, {}, solidus_server.logger);
-      if (!resource.resource.url) return res.json(400, {error: "Invalid 'url' parameter"});
+      var cached_resource = new CachedResource(req.query.url, solidus_server.auth, {}, solidus_server.logger);
+      if (!cached_resource.resource.url) return res.json(400, {error: "Invalid 'url' parameter"});
 
-      resource.get(function(err, cached_resource) {
+      cached_resource.get(function(err, resource_response) {
         if (err) return res.json(400, {error: err});
 
         if (!self.options.dev) {
           res.set({
-            'Cache-Control': 'public, max-age=' + cached_resource.maxAge(),
-            'Expires': new Date(cached_resource.expires_at).toUTCString()
+            'Cache-Control': 'public, max-age=' + resource_response.maxAge(),
+            'Expires': new Date(resource_response.expires_at).toUTCString()
           });
         }
-        res.json(cached_resource.data);
+        res.json(resource_response.data);
       });
     });
   };
